requirements.txt
click
dicetray
discord.py
